requirements.txt
python-dotenv==1.1.0
openai==1.97.0
torch==2.7.0
torchaudio==2.7.0
silero==0.4.1

livekit==1.0.11
livekit-agents==1.2.1
livekit-plugins-openai==1.2.1
livekit-plugins-silero==1.2.1
livekit-plugins-turn-detector==1.0.17
